# 【PPDiffusers】增加FD ControlNet text2img, img2img, inpaint_legacy, hiresfix pipeline