--- a/ppdiffusers/deploy/controlnet/infer.py
+++ b/ppdiffusers/deploy/controlnet/infer.py
@@ -88,7 +88,7 @@ def parse_arguments():
     parser.add_argument(
         "--parse_prompt_type",
         type=str,
-        default="raw",
+        default="lpw",
         choices=[
             "raw",
             "lpw",