--- a/ppdiffusers/deploy/controlnet/infer_dygraph.py
+++ b/ppdiffusers/deploy/controlnet/infer_dygraph.py
@@ -74,7 +74,7 @@ def parse_arguments():
     parser.add_argument(
         "--parse_prompt_type",
         type=str,
-        default="raw",
+        default="lpw",
         choices=[
             "raw",
             "lpw",