--- a/ppdiffusers/deploy/infer_dygraph.py
+++ b/ppdiffusers/deploy/infer_dygraph.py
@@ -57,7 +57,7 @@ def parse_arguments():
     parser.add_argument(
         "--parse_prompt_type",
         type=str,
-        default="raw",
+        default="lpw",
         choices=[
             "raw",
             "lpw",