--- a/ppdiffusers/deploy/requirements.txt
+++ b/ppdiffusers/deploy/requirements.txt
@@ -1,0 +1,1 @@
+ppdiffusers>=0.16.1