… ppdiffusers/deploy/README.md
… ppdiffusers/deploy/controlnet/README.md
… ppdiffusers/deploy/controlnet/controlnet_infer.py
… ppdiffusers/deploy/scripts/test_controlnet_infer_dygraph.sh
… ppdiffusers/deploy/scripts/test_controlnet_infer_fd.sh
… ppdiffusers/deploy/scripts/test_infer.sh
… ppdiffusers/deploy/scripts/test_infer_dygraph.sh
… ppdiffusers/deploy/scripts/test_infer_fd.sh
… ppdiffusers/examples/community/pipeline_fastdeploy_stable_diffusion_hires_fix.py
… ppdiffusers/examples/community/reference_only.py
… ppdiffusers/examples/community/stable_diffusion_controlnet_img2img.py
… ppdiffusers/examples/community/stable_diffusion_hires_fix.py
… ppdiffusers/examples/community/stable_diffusion_mega.py
… ppdiffusers/ppdiffusers/pipelines/stable_diffusion/pipeline_fastdeploy_stable_diffusion_img2img.py
… ppdiffusers/ppdiffusers/pipelines/stable_diffusion/pipeline_fastdeploy_stable_diffusion_inpaint.py
… ppdiffusers/ppdiffusers/pipelines/stable_diffusion/pipeline_fastdeploy_stable_diffusion_inpaint_legacy.py
… ppdiffusers/ppdiffusers/pipelines/stable_diffusion/pipeline_fastdeploy_stable_diffusion_mega.py
… ppdiffusers/ppdiffusers/pipelines/stable_diffusion/pipeline_fastdeploy_stable_diffusion_upscale.py
… ppdiffusers/ppdiffusers/pipelines/stable_diffusion/pipeline_stable_diffusion_inpaint_legacy.py
… ppdiffusers/ppdiffusers/schedulers/preconfig/preconfig_scheduling_euler_ancestral_discrete.py